Initial setup for awilix & file-location changes. moved the index.js file to the presentation layer. Added awilix to package.json Created a currently empty container.js file for dependency injections
