+++ package.json
@@ -12,6 +12,7 @@
   "type": "module",
   "dependencies": {
     "@google-cloud/vision": "^3.1.2",
+    "awilix": "^8.0.0",
     "blob": "^0.1.0",
     "body-parser": "^1.20.2",
     "express": "^4.18.2",
