+++ presentation/routers/imageRouter.mjs
